crawler/requirements.txt
selenium>=4.1.0
scrapy-selenium>=0.0.7
scrapy>=2.11.2
requests>=2.28.2
neo4j>=5.20.0
textblob>=0.18.0.post0
nltk>=3.8.1
gensim==4.3.2
scipy==1.10.1
requests==2.32.3
python-dotenv>=1.0.1